pyproject.toml
[project]
name = "colorclade"
version = "0.0.0"
description = "colorclade draws phylogenies with hierarchical coloring for easier visual comparison."
authors = [
    { name = "Matthew Andres Moreno", email = "m.more500@gmail.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
license = {text = "MIT license"}
dependencies = [
  "alifedata_phyloinformatics_convert",
  "Biopython",
  "matplotlib",
  "numpy",
  "opytional",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
homepage = "https://github.com/mmore500/colorclade"
documentation = "https://github.com/mmore500/colorclade"
repository = "https://github.com/mmore500/colorclade"
tracker = "https://github.com/mmore500/colorclade/issues"

[project.optional-dependencies]
dev = [
    "black==22.10.0",
    "bump2version==1.0.1",
    "isort==5.12.0",
    "pip-tools==7.3.0",
    "pytest==6.2.5",
    "seaborn==0.13.0",
    "ruff==0.1.11",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]
namespaces = false

# adapted from https://tox.wiki/en/latest/example/package.html
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "cython>=0.29",
  "setuptools >= 35.0.2",
  "setuptools_scm >= 2.0.0, <3",
]

[tool.black]
line_length = 79

[tool.isort]
force_sort_within_sections = true
line_length = 79
profile = "black"
skip_gitignore = true
